Change how the Dart library name is generated
The package specifier from the .proto file is now used to set the
library name. In the case where there is no package specifier the
previous method of using the file name is used except that the first
letter is not changed into upper-case.

R=sigmund@google.com
BUG=https://github.com/dart-lang/dart-protoc-plugin/issues/7

Review URL: https://chromiumcodereview.appspot.com//23522040