Use package names as import prefixes when generating code
The generated code does not include the package name in the generated
top-level classes. If several packages have the same class this causes
conflicts. This change will add a prefix for each import in the
generated file to solve this. The prefix used is the package name.

R=sigmund@google.com
BUG=https://github.com/dart-lang/dart-protobuf/issues/12

Review URL: https://chromiumcodereview.appspot.com//93743006