Fix issue with importing packages where the package name included a dot
The generated code for imports where just using the package name as
the prefix. This did not work if the package name included a dot. The
prefix is not generated by replacing dots with the dollar
character. The dollar character is valid in Dart identifiers but not
in protocol buffer package names so no name clashes should happen.

R=sigmund@google.com

Review URL: https://chromiumcodereview.appspot.com//128003003