Add an option to extend Map in GeneratedMessage subclasses
When enabled, the GeneratedMessage acts as a map with a fixed number
of (key, value) pairs, one for each field. The key is the field
name (string), using same name as the Dart getter, not the original
field name in the .proto file. The value of each pair starts out with
the field's default value. Extensions and unknown fields aren't
included in the map.

Unlike regular maps, pairs can't be added or removed. Values can be
changed using map operations only for non-repeated fields, where the
new value has a matching type.

The following operations don't behave like a regular map:

 - clear() sets fields back to their default values, rather
than removing all pairs in the map.

 - operator == only returns true if the other map is also a
GeneratedMessage with the same type. It does a deep equality check
that compares all fields in the message, including extensions and
unknown fields. As a result, comparisons using == aren't commutative
between a GeneratedMessage and other kinds of maps.

- hashCode is consistent with operator ==. A GeneratedMessage won't
have the same hashCode as a regular map containing the same pairs.

This option is experimental and may change without notice.
But if you really want to try it out, see map_api.proto and
dart_options.proto for how to turn it on.

BUG=https://github.com/dart-lang/dart-protoc-plugin/issues/44