fix inconsistent firing of events for extension fields
Before we would fire events for some extension field
mutations but not others. Change it to fire no events
for these fields for now.

(Using _setField consistently should make it easier to
send events later if we decide to do that.)

Also, remove the optional third argument to setField().
It shouldn't be part of the public API since it can
be used to defeat field validation.

BUG=
R=sgjesse@google.com

Review URL: https://chromiumcodereview.appspot.com//1282103003.