Write a better benchmark framework.
We can take samples while showing progress on both the VM and in a browser.
Benchmark reports can be saved and loaded as a proto buffer (this is
by copying and pasting the JSON for now).
The reports include the environment where the benchmark was run.

Includes some baseline data for 0.4.2, for comparison.

Todo:
  - add information from pubspec.lock
  - add a view showing the diffs between two benchmark reports

Also fixes a bug:
  - The non-repeated field results were bogus.
    (It was running the repeated benchmark!)

BUG=
R=sgjesse@google.com

Review URL: https://chromiumcodereview.appspot.com//1314823009.