Create _extensions and unknownFields lazily
These fields are often unused. We only need to create
them when there is an extension field or an unknown field.

This speeds up the readInts benchmark by about 30% when
running on the Dart VM. There is no difference as far
as I can tell when using dart2js.

BUG=
R=sgjesse@google.com

Review URL: https://chromiumcodereview.appspot.com//1327213002.