fix bug in property benchmarks and improve accuracy
The getter benchmarks had a bug where most of the test could be skipped,
without throwing an exception. As a result, the measurements were
much higher than they should be for getters where the property
is not set and we fall back on the default.

Also, increase the number of repetitions to reduce overhead a bit.

Reran the property benchmarks.

BUG=
R=sgjesse@google.com

Review URL: https://chromiumcodereview.appspot.com//1378463002.