Generate separate .pbjson.dart files for constants
This is to reduce the amount of code parsed in Dartium
for the common case where a .proto doesn't define any services.
(Tree shaking does not help us in this case.)

Service definitions still need to import the JSON constants.
But it might also help a bit in this case, since the
dependencies between files are more fine-grained than before.

BUG=
R=sgjesse@google.com

Review URL: https://chromiumcodereview.appspot.com//1941153002 .