Change how to set the Dart name of a field
The old way was by passing a --field_name flag to the protoc
generator. (However, it's doubtful whether anyone was using it.)

The new way is to use the field_name option on the .proto file.
For example:

  optional string some_field = 1 [
    (dart_options.dart_name) = "renamedField"
  ];

Based on a patch by Frederik Mutzel.

BUG=
R=frederikmutzel@google.com, sgjesse@google.com

Review URL: https://chromiumcodereview.appspot.com//2043913005 .