Always serialize 64-bit integers as strings in JSON.
This improves compatibility with some other implementations that only
expect strings. Existing implementations using the previous format
should handle both strings and numbers, so switching to pure strings
should be safe.

There is a slight performance impact to this, but only really noticeable
on messages with 100's of repeated int64 fields, so we decided against
adding an option for this.

BUG=https://github.com/dart-lang/dart-protobuf/issues/58
R=skybrian@google.com

Review-Url: https://codereview.chromium.org//2612233003 .