Don't use readlink in bin/protoc-gen-dart. (#78)
`readlink -f` doesn't work on macOS, unfortunately. The main side-effect of this change is that making a symlink to `bin/protoc-gen-dart` won't work anymore, since it won't correctly resolve it's directory.

The preferred way to run the plugin is using `pub global activate` or explicitly passing the path on the protoc commmand line, so this is mainly for development.

Fixes #77.