Fix unnecessary_new lint in generated code.
I also added this as a lint check for protoc_plugin because the easiest
thing to fix this was to run a sed -i 's/new //g' on all files.

I checked it manually but worst case scenario there's a missing "new"
word on a comment.

Fixes one of the warnings from #135