Use FieldDescriptorProto.json_name for representing field names.
This seems to be the right thing to do if we want to emit
proto3 JSON.
Currently we were using the dart mangled name, that is also camel-case
but has a number of dart-specific disambiguations applied making it unsuitable for
interop.

If we want to parse JSON we need also know the field_name (with
underscores). That is for a later PR...