Move well-known-type support code to mixins.
This makes it easier to implement support for additional well-known types
as it moves more code away from being injected strings.

The static methods still have to be injected, because you cannot mix in
static methods.

Changes mixin imports in generated code to use a prefix instead of a name.

Also changes the import of dart:core to not use a `show` list. It was
not needed after we use prefix imports.

Also moves the protobuf/lib/src/mixins_meta.dart and associated tests to
protoc_plugin/lib/mixins.dart. It was not used there.