Remove unnecessary dart:async import.
I'm trying to enforce this inside google3, and so it would be nice to be
able to land this.

This requires Dart 2.1, so this is technically a breaking change. If
that's not OK, then I can abandon this PR and make an exception inside
google3 instead.

Note that Dart 2.1 has been out for quite a long time now.