Restore possibility to use reflective API to create default field values using getField()
We do not actually have to prevent/throw when trying to create a default
value for a PbMap field.

Initially when working on removing BuilderInfo references it was
necessary to remove default creation of map fields due to not
having the [BuilderInfo] object for the map which was required for the
PbMap constructor.

Though a later iteration of that CL removed the 3rd parameter to PbMap
(or rather made that parameter optional and ignored):

  -  PbMap(this.keyFieldType, this.valueFieldType, this._entryBuilderInfo)

  +  // The provided [info] will be ignored.
  +  PbMap(this.keyFieldType, this.valueFieldType, [BuilderInfo? info])

So we can restore `defaultOrMake()`.