Testing: don't return static methods in findMemberNames (#618)
This refactors the fix for #527 and improves on it. We now filter out static
methods in `findMemberNames`. Static methods cannot be overridden or shadowed
in subclasses so we don't consider them as reserved.

Static methods `GeneratedMessage.$_defaultFor` and `ProtobufEnum.initByValue`
are removed from reserved names lists. These were probably added because we
didn't properly remove static methods in `findMemberNames` before.

Closes #527