Update lints
This updates the lints used in the library and plugin to make it easier
to keep the code in sync with the internal version.

The only change is we now enable `omit_local_variable_types`.

This lint was enabled before in 52544ea (and then disabled in 45e054a),
so most of the code, other than a few tests, already follow this lint.
So no changes needed except a few lines in tests.

Lints `control_flow_in_finally` and `prefer_spread_collections` are
removed as they're already in recommended lints and enabled.