Revert contructor arguments
This PR reverts #442 (7ac17d0f)

This patch was ported to internal in cl/449707130, but benchmarking
revealed that the new constructors add to release binary sizes even when
no arguments are passed to the constructors.

To keep the two code bases in sync I'm reverting this change.

These new constructor arguments were never announced in the changelog,
but I still added an entry to show how to migrate code that uses them.