Make FieldSet lists use the same representation when empty and otherwise
I think `List.filled` could return different `List`s when length is 0
and otherwise, but I checked SDK code and at least on JS that's not the
case.

With `[]` syntax the list will be modifiable, and representation can be
different than `List.filled(..., growable: false)`.

Updated `_zeroList` documentation for clarity.