Fix presence of bytes fields
`_FieldSet._$has` previously returned `true` if a list field is empty.

Because `bytes` fields are represented as `List<int>`, this caused
incorrect presence when we set a bytes field to `[]`.

Remove the special case for lists in `_$has` to fix presence checks for
`bytes` fields.

The special case for lists in `_$has` is not necessary for repeated
fields as we don't generate hazzers for repeated fields.

Fixes #690