Relax a type cast in `CodedBufferWriter.writeField`
In protobuf library we use `PbList` for all repeated fields.

However the method `CodedBufferWriter.writeField` is public, and we have
users that call it with `List` types other than `PbList`. (and with
`type` values copied from protobuf library source code)

To avoid breakage, cast the `dynamic` value to `List` instead of
`PbList`.