Fix initializing required fields with proto3 JSON null values
[proto3 JSON spec][1] says this about `null` values:

> If a value is missing in the JSON-encoded data or if its value is
> null, it will be interpreted as the appropriate default value when
> parsed into a protocol buffer

Normally empty field set represents the default message, but for that
representation to work with proto2 required fields, serializers need to
check for required fields in the `BuilderInfo` and for fields that are
not in the field set serialize the default value for the field.

Currently we don't do this, which results in generating messages with
missing required fields. See #760 for a repro.

Checking missing required fields in serializers should fix the issue,
but as a simpler fix, this PR deserializes proto3 JSON `null` values to
field sets with required fields set to defaults. This way serializers do
not need to be tweaked.

Fixes #760

[1]: https://developers.google.com/protocol-buffers/docs/proto3#json