Add methods that construct read-only messages
This patch adds the following methods to `GeneratedMessage`:
* `mergeFromBufferFrozen`
* `mergeFromCodedBufferReaderFrozen`
* `mergeFromJsonFrozen`
* `mergeFromJsonMapFrozen`

All of them add to the current message by deserializing the provided
data and mark the new parts of the message (including the message
itself) as read-only.

Note that sub-messages that are already part of the current message are
not marked as read-only.

Closes #627