Remove event support
This removes `PbEventMixin`, `PbFieldChange`, and `EventBuffer` types,
and the event plugin mixin support from the plugin.

This mixin is currently not used internally and we don't want to
introduce uses to it to keep the API surface small, to make it easier to
make changes or migrate to another protobuf library.

Fixes #738.

cl/571893384