Keep a `ByteData` around in `CodedBufferReader` to avoid repeated `ByteData` allocs
In Dart, the only way to convert a list of bytes to an `int` or `double` is by
using a `ByteData`.

Currently `CodedBufferReader` allocates a new `ByteData` for every fixed-size
`int` or `float` parsing.

With this change we now allocate a `ByteData` once initialization and reuse it.

Benchmark results, using the same benchmark reported in #888:

|                              | Before     | After      | Diff                |
|------------------------------|------------|------------|---------------------|
| AOT                          |  26,763 us |  25,305 us | - 1,458 us, - 5.4%  |
| JIT                          |  24,573 us |  25,521 us | +   948 us, + 3.8%  |
| dart2js -O4                  | 212,000 us | 199,300 us | -12,700 us, - 5.9%  |
| dart2wasm --omit-type-checks |  89,750 us |  78,958 us | -10,792 us, -12.0%  |