Make `CodedBufferWriter.writeRawByte` argument type more accurate
`writeRawBytes` adds bytes to the output and it will only be fast when
the argument is `Uint8List`.

Since all the users already pass `Uint8List`, update the argument type.

This allows simplifyping `_copyInto`. Because it has only one call site,
we simplify it and inline it in the call site.

cl/576427817