Update getList and getMap return types (#903)
This improves iteration performance by making `moveNext` and `current` direct
calls in kernel. #902 does not have the same effect on `moveNext` and `current`
calls, this change is needed even with #902.

This change was not possible before #880 as users could override the list and
map types to types that are not `PbList`s or `PbMap`s.

Note: changes in the message field types (the plugin changes in this PR) are
not necessary for the kernel improvements, but it doesn't hurt to generate more
precise types everywhere.