Replace removed `dart:typed_data` type in test code
`Unmodifiable...View` types were deprecated in favor of
`asModifiableView`s members introduced in SDK 3.3.0.

These types are now removed, so the protobuf package tests can't be
built with the most recent SDK.

Bump the SDK dependency to 3.3.0 (the version introduced
`asModifiableView`) and replace the removed type with `asModifiableView`
in the test.