Update SDK version to 3.7, reformat with 3.7 formatter (#1024)
Internally we're at SDK 3.7. Because it's difficult to make changes *without*
formatting in g3, contributors create CLs with accidental formatting fixes.

In cl/781448639 we format the internal code, and this PR updates to the same
SDK version with g3 and formats the code, making the internal and external
versions of the code meet at the same SDK version and formatting settings.

Manual changes in this PR:

- Update pubspec.yaml files:
  - Bump minor versions
  - Bump SDK versions

- Update CHANGELOGs

- Update dart_style used in protoc_plugin golden tests

- Regenerate mono_repo setup with `mono_repo generate`

Rest of the changes are generated by `dart format`.