Split protobuf package into libraries (#1026)
This is the first part of the PRs that sync internal JSON decoding chages.

To be able to conditionally import different JSON decoders, this splits the
monolithic protobuf package into libraries.

These parts are now internal libraries:

- consts.dart
- json_parsing_context.dart
- permissive_compare.dart
- type_registry.dart
- utils.dart

These changes are not useful on their own, they're to prepare the library for
the rest of the PRs and to keep the PRs small and easier to review.