CI: Run all of the checks on every change
Drop paths in workflow specs to run all of the checks on every change.

Dependencies between directories in this project is a bit tricky.
protoc_plugin uses protobuf, which is already incorrect in the paths
currently. Benchmarks also depend on both the plugin and the library.

Instead of manually specifying and maintaining tricky dependencies just
run all of the checks on all changes. No need to optimize this and risk
missing checks until CI time becomes a problem.