Fix JS interop annotations
In a declaration like

    @JS('JSON')
    extension type _JSON._(JSObject _) implements JSObject {
      @JS('JSON.stringify')
      external static JSString _stringify(JSObject value);
    }

When I call `_JSON._stringify(...)`, dart2js calls `JSON.stringify`, but
dart2wasm calls `JSON.JSON.stringify`, which is wrong.

Reading [1] it looks like the annotation `JSON.stringify` should just be
`stringify`. If I do that the code works with both dart2wasm and
dart2js.

(I've reported dart2wasm's handling of the annotations above and we'll
work on a fix for it on dart2wasm.)

[1]: https://dart.dev/interop/js-interop/usage#js