Use a map to cache the `valueOf` functions for enums (#1047)
(This is cl/658131451 ported to open source, not my CL)

CL description: (slightly edited)

This is a generated-code size optimization.

The main change is to generate the valueOf functions for protobuf enums
programatically, rather than (1) use the per-enum `FooPbEnum.valueOf` static
function tear-foo which (2) uses a per-enum static variable that is lazily
initialized to map from int value to the protobuf enum value.

With ~1000 protobuf enums, this allows most of the `valueOf` static function
tear-off closures to tree-shaken, together with their functions and the code
for the initializer of the map.

The secondary change is to use more specialized 'add' methods to `BuilderInfo`,
allowing many fields to be 'defined' with a call with fewer arguments.

Together, these changes reduce the size of the main unit of an app by about
1.3%, and would have a similar effect on native targets.

This change requires
https://github.com/dart-lang/sdk/commit/175dc05db3cd626874c17ceb03336065a03375bb,
which was released with Dart 3.6.0.