Improve golden file checking
Currently when I have a typo in the golden file path, `expectGolden`
creates a new file, which makes the test pass.

Update `expectGolden` to only create (or update if already exists) a
golden file when the env variable `PROTOC_UPDATE_GOLDENS` is set.

Without the env variable it expects the file to exist, with the right
contents. So typos now cause test failures as expected.