Improve golden file checking (#1053)
Currently when I have a typo in the golden file path, `expectGolden` creates a
new file, which makes the test pass.

Locally I can notice the new file if I run `git status`, but if I'm not careful
and rely on CI, because the tests pass, I can merge the PR.

Update `expectGolden` to only create (or update if already exists) a golden
file when the env variable `PROTOC_UPDATE_GOLDENS` is set.

Without the env variable it expects the file to exist, with the right contents.
So typos now cause test failures as expected.