Run plugin with asserts when building test protos (#1055)
Currently we run the tests with assertions but that doesn't cover some of the
code paths.

When building the test protos we don't run the plugin with assertions.

Add a new executable `protoc-gen-dart-debug` that is the same as
`protoc-gen-dart`, but passes `--enable-asserts` to Dart. Use it when building
test protos.

This triggers an assertion which was reported in #608. The assertion is clearly
wrong (we call the function for repeated fields a few lines above the
assertion) and no one knows or remembers why it's there. Remove the assertion.

Closes #608.