Generalize argument type of `PbList.from` from `List<T>` to `Iterable<T>` (#1054)
This is more consistent with `List.from` and also more efficient as we can
avoid redundant `List` allocations when we have an `Iterable` instead of a
`List`. This commonly happens when we create a `PbList` from a chain of `map`,
`where` etc. calls.

Also update the the factory body to use `List.of` instead of `List.from` as we
know the `Iterable`'s element type. (`List.from` takes `Iterable<dynamic>`,
`List.of` takes `Iterable<T>`)

cl/810421932