Fix a few extension field bugs (#1062)
When an extension field set isn't initialized before freezing the message:

- `addExtension` allows adding and modifying extensions.
- `getExtension` returns non-frozen defaults that can be modified. 

Fix this by initializing the extension field set as frozen when the parent
message is frozen.

Secondly, extension field set returns differently typed lists when it's frozen
and not frozen. Fix this by always creating the list via the field's
`FieldInfo`.

Also refactor the code a little bit: `GeneratedMessage` by-passes a layer of
abstraction and accesses unknown field sets directly. Instead make it call
`FieldSet` always.

New `FieldSet` methods are added as public (without an underscore in the name)
to make it clear that they're a public interface (can be called from
`GeneratedMessage`).

(Ideally `FieldSet` should be a separate library, but that requires a lot of
refactoring to split the parts into libraries.)

Also remove duplicate "is repeated" and "ensure writable" checks when modifying
extension fields.

---

cl/819146715